XGBoost-predictor 0.3.5 Maven central dependency added. Removed other unused dependencies. Moved common code to util classes. Moved new code to the end of existing classes. Removed leftovers/unused code. Optimized domain variables access during graph conversion.
