+++ h2o-extensions/xgboost/src/main/java/hex/tree/xgboost/XGBoostModel.java
@@ -28,7 +28,6 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
 import static hex.tree.xgboost.XGBoost.makeDataInfo;
 import static hex.genmodel.algos.xgboost.XGBoostMojoModel.ObjectiveType;
@@ -38,125 +37,6 @@ public class XGBoostModel extends Model<XGBoostModel, XGBoostModel.XGBoostParame
   private XGBoostModelInfo model_info;
 
   public XGBoostModelInfo model_info() { return model_info; }
-
-  @Override
-  public SharedTreeGraph convert(final int treeNumber, final String treeClassName) {
-    GradBooster booster = null;
-    try {
-      booster = new Predictor(new ByteArrayInputStream(model_info._boosterBytes)).getBooster();
-    } catch (IOException e) {
-      Log.err(e);
-      H2O.fail(e.getMessage());
-    }
-
-    if (!(booster instanceof GBTree)) {
-      throw new IllegalArgumentException(String.format("Given XGBoost model is not backed by a tree-based booster. Booster class is %d",
-              booster.getClass().getCanonicalName()));
-    }
-
-    final RegTree[][] groupedTrees = ((GBTree) booster).getGroupedTrees();
-    final int treeClass = getXGBoostClassIndex(treeClassName);
-    if (treeClass >= groupedTrees.length) {
-      throw new IllegalArgumentException(String.format("Given XGBoost model does not have given class '%s'.", treeClassName));
-    }
-
-    final RegTree[] treesInGroup = groupedTrees[treeClass];
-
-    if (treeNumber >= treesInGroup.length || treeNumber < 0) {
-      throw new IllegalArgumentException(String.format("There is no such tree number for given class. Total number of trees is %d.", treesInGroup.length));
-    }
-
-    final RegTreeNode[] treeNodes = treesInGroup[treeNumber].getNodes();
-    assert treeNodes.length >= 1;
-
-    SharedTreeGraph sharedTreeGraph = new SharedTreeGraph();
-    final SharedTreeSubgraph sharedTreeSubgraph = sharedTreeGraph.makeSubgraph(_output._training_metrics._description);
-    treeNodes[0].split_index();
-
-    final FeatureProperties featureProperties = assembleFeatureNames(model_info._dataInfoKey.get()); // XGBoost's usage of one-hot encoding assumed
-    constructSubgraph(treeNodes, sharedTreeSubgraph.makeRootNode(), 0, sharedTreeSubgraph, featureProperties, true); // Root node is at index 0
-    return sharedTreeGraph;
-  }
-
-  private static void constructSubgraph(final RegTreeNode[] xgBoostNodes, final SharedTreeNode sharedTreeNode,
-                                        final int nodeIndex, final SharedTreeSubgraph sharedTreeSubgraph,
-                                        final FeatureProperties featureProperties, boolean inclusiveNA) {
-    final RegTreeNode xgBoostNode = xgBoostNodes[nodeIndex];
-    // Not testing for NaNs, as SharedTreeNode uses NaNs as default values.
-    //No domain set, as the structure mimics XGBoost's tree, which is numeric-only
-    if (featureProperties._oneHotEncoded[xgBoostNode.split_index()]) {
-      //Shared tree model uses < to the left and >= to the right. Transforiming one-hot encoded categoricals
-      // from 0 to 1 makes it fit the current split description logic
-      sharedTreeNode.setSplitValue(1.0F);
-    } else {
-      sharedTreeNode.setSplitValue(xgBoostNode.getSplitCondition());
-    }
-    sharedTreeNode.setPredValue(xgBoostNode.getLeafValue());
-    sharedTreeNode.setCol(xgBoostNode.split_index(), featureProperties._names[xgBoostNode.split_index()]);
-    sharedTreeNode.setInclusiveNa(inclusiveNA);
-    sharedTreeNode.setNodeNumber(nodeIndex);
-
-    if (xgBoostNode.getLeftChildIndex() != -1) {
-      constructSubgraph(xgBoostNodes, sharedTreeSubgraph.makeLeftChildNode(sharedTreeNode),
-              xgBoostNode.getLeftChildIndex(), sharedTreeSubgraph, featureProperties, xgBoostNode.default_left());
-    }
-
-    if (xgBoostNode.getRightChildIndex() != -1) {
-      constructSubgraph(xgBoostNodes, sharedTreeSubgraph.makeRightChildNode(sharedTreeNode),
-              xgBoostNode.getRightChildIndex(), sharedTreeSubgraph, featureProperties, !xgBoostNode.default_left());
-    }
-  }
-
-
-  private static FeatureProperties assembleFeatureNames(final DataInfo di) {
-    String[] coefnames = di.coefNames();
-    assert (coefnames.length == di.fullN());
-    int numCatCols = di._catOffsets[di._catOffsets.length - 1];
-
-    String[] featureNames = new String[di.fullN()];
-    boolean[] oneHotEncoded = new boolean[di.fullN()];
-    for (int i = 0; i < di.fullN(); ++i) {
-      featureNames[i] = coefnames[i];
-      if (i < numCatCols) {
-        oneHotEncoded[i] = true;
-      }
-    }
-
-    return new FeatureProperties(featureNames, oneHotEncoded);
-  }
-
-  private static class FeatureProperties {
-    private String[] _names;
-    private boolean[] _oneHotEncoded;
-
-    public FeatureProperties(String[] names, boolean[] oneHotEncoded) {
-      _names = names;
-      _oneHotEncoded = oneHotEncoded;
-    }
-  }
-
-  private final int getXGBoostClassIndex(final String treeClass) {
-    final ModelCategory modelCategory = _output.getModelCategory();
-    if (treeClass == null && ModelCategory.Regression.equals(modelCategory)) return 0;
-    if (treeClass == null && !ModelCategory.Regression.equals(modelCategory)) {
-      throw new IllegalArgumentException("Non-regressional models require tree class specified.");
-    }
-
-    final String[] domain = Objects.requireNonNull(model_info._dataInfoKey.get())
-            ._adaptedFrame.vec(_parms._response_column).domain();
-    if(ModelCategory.Regression.equals(modelCategory) && treeClass != null){
-      throw new IllegalArgumentException("There should be no tree class specified for regression.");
-    }
-    final int treeClassIndex = ArrayUtils.find(domain, treeClass);
-
-    if (ModelCategory.Binomial.equals(modelCategory) && treeClassIndex != 0) {
-      throw new IllegalArgumentException(String.format("For binomial XGBoost model, only one tree for class %s has been built.", domain[0]));
-    } else if (treeClassIndex < 0) {
-      throw new IllegalArgumentException(String.format("No such class '%s' in tree.", treeClass));
-    }
-
-    return treeClassIndex;
-  }
 
   public static class XGBoostParameters extends Model.Parameters {
     public enum TreeMethod {
@@ -701,6 +581,96 @@ public class XGBoostModel extends Model<XGBoostModel, XGBoostModel.XGBoostParame
     if (model_info()._dataInfoKey !=null)
       model_info()._dataInfoKey.get().remove(fs);
     return super.remove_impl(fs);
+  }
+
+  @Override
+  public SharedTreeGraph convert(final int treeNumber, final String treeClassName) {
+    GradBooster booster = null;
+    try {
+      booster = new Predictor(new ByteArrayInputStream(model_info._boosterBytes)).getBooster();
+    } catch (IOException e) {
+      Log.err(e);
+      throw new IllegalStateException("Booster bytes inaccessible. Not able to extract the predictor and construct tree graph.");
+    }
+
+    if (!(booster instanceof GBTree)) {
+      throw new IllegalArgumentException(String.format("Given XGBoost model is not backed by a tree-based booster. Booster class is %d",
+              booster.getClass().getCanonicalName()));
+    }
+
+    final RegTree[][] groupedTrees = ((GBTree) booster).getGroupedTrees();
+    final int treeClass = getXGBoostClassIndex(treeClassName);
+    if (treeClass >= groupedTrees.length) {
+      throw new IllegalArgumentException(String.format("Given XGBoost model does not have given class '%s'.", treeClassName));
+    }
+
+    final RegTree[] treesInGroup = groupedTrees[treeClass];
+
+    if (treeNumber >= treesInGroup.length || treeNumber < 0) {
+      throw new IllegalArgumentException(String.format("There is no such tree number for given class. Total number of trees is %d.", treesInGroup.length));
+    }
+
+    final RegTreeNode[] treeNodes = treesInGroup[treeNumber].getNodes();
+    assert treeNodes.length >= 1;
+
+    SharedTreeGraph sharedTreeGraph = new SharedTreeGraph();
+    final SharedTreeSubgraph sharedTreeSubgraph = sharedTreeGraph.makeSubgraph(_output._training_metrics._description);
+
+    final XGBoostUtils.FeatureProperties featureProperties = XGBoostUtils.assembleFeatureNames(model_info._dataInfoKey.get()); // XGBoost's usage of one-hot encoding assumed
+    constructSubgraph(treeNodes, sharedTreeSubgraph.makeRootNode(), 0, sharedTreeSubgraph, featureProperties, true); // Root node is at index 0
+    return sharedTreeGraph;
+  }
+
+  private static void constructSubgraph(final RegTreeNode[] xgBoostNodes, final SharedTreeNode sharedTreeNode,
+                                        final int nodeIndex, final SharedTreeSubgraph sharedTreeSubgraph,
+                                        final XGBoostUtils.FeatureProperties featureProperties, boolean inclusiveNA) {
+    final RegTreeNode xgBoostNode = xgBoostNodes[nodeIndex];
+    // Not testing for NaNs, as SharedTreeNode uses NaNs as default values.
+    //No domain set, as the structure mimics XGBoost's tree, which is numeric-only
+    if (featureProperties._oneHotEncoded[xgBoostNode.split_index()]) {
+      //Shared tree model uses < to the left and >= to the right. Transforiming one-hot encoded categoricals
+      // from 0 to 1 makes it fit the current split description logic
+      sharedTreeNode.setSplitValue(1.0F);
+    } else {
+      sharedTreeNode.setSplitValue(xgBoostNode.getSplitCondition());
+    }
+    sharedTreeNode.setPredValue(xgBoostNode.getLeafValue());
+    sharedTreeNode.setCol(xgBoostNode.split_index(), featureProperties._names[xgBoostNode.split_index()]);
+    sharedTreeNode.setInclusiveNa(inclusiveNA);
+    sharedTreeNode.setNodeNumber(nodeIndex);
+
+    if (xgBoostNode.getLeftChildIndex() != -1) {
+      constructSubgraph(xgBoostNodes, sharedTreeSubgraph.makeLeftChildNode(sharedTreeNode),
+              xgBoostNode.getLeftChildIndex(), sharedTreeSubgraph, featureProperties, xgBoostNode.default_left());
+    }
+
+    if (xgBoostNode.getRightChildIndex() != -1) {
+      constructSubgraph(xgBoostNodes, sharedTreeSubgraph.makeRightChildNode(sharedTreeNode),
+              xgBoostNode.getRightChildIndex(), sharedTreeSubgraph, featureProperties, !xgBoostNode.default_left());
+    }
+  }
+
+
+  private final int getXGBoostClassIndex(final String treeClass) {
+    final ModelCategory modelCategory = _output.getModelCategory();
+    if (treeClass == null && ModelCategory.Regression.equals(modelCategory)) return 0;
+    if (treeClass == null && !ModelCategory.Regression.equals(modelCategory)) {
+      throw new IllegalArgumentException("Non-regressional models require tree class specified.");
+    }
+
+    final String[] domain = _output._domains[_output._domains.length - 1];
+    if(ModelCategory.Regression.equals(modelCategory) && treeClass != null){
+      throw new IllegalArgumentException("There should be no tree class specified for regression.");
+    }
+    final int treeClassIndex = ArrayUtils.find(domain, treeClass);
+
+    if (ModelCategory.Binomial.equals(modelCategory) && treeClassIndex != 0) {
+      throw new IllegalArgumentException(String.format("For binomial XGBoost model, only one tree for class %s has been built.", domain[0]));
+    } else if (treeClassIndex < 0) {
+      throw new IllegalArgumentException(String.format("No such class '%s' in tree.", treeClass));
+    }
+
+    return treeClassIndex;
   }
 
 }
