+++ h2o-extensions/xgboost/src/main/java/ml/dmlc/xgboost4j/java/XGBoostRegTree.java
@@ -2,6 +2,7 @@ package ml.dmlc.xgboost4j.java;
 
 import biz.k11i.xgboost.tree.RegTree;
 import biz.k11i.xgboost.tree.RegTreeImpl;
+import biz.k11i.xgboost.tree.RegTreeNode;
 import biz.k11i.xgboost.util.FVec;
 import biz.k11i.xgboost.util.ModelReader;
 import water.util.UnsafeUtils;
@@ -73,9 +74,10 @@ public class XGBoostRegTree implements RegTree {
   }
 
   @Override
-  public RegTreeImpl.Node[] getNodes() {
+  public RegTreeNode[] getNodes() {
     throw new UnsupportedOperationException();
   }
+
 
   private static int readNumNodes(ModelReader reader) throws IOException {
     int numRoots = reader.readInt();
