+++ h2o-genmodel-extensions/xgboost/build.gradle
@@ -5,7 +5,9 @@ description = "H2O GenModel XGBoost support"
 dependencies {
     compile project(":h2o-genmodel")
 
-    compile files('/home/pavel/wrk/xgboost-predictor/xgboost-predictor/build/libs/xgboost-predictor-0.3.4.jar');
+    compile('ai.h2o:xgboost-predictor:0.3.5') {
+        exclude group: 'net.jafama', 'module': 'jafama'
+    }
 
     // XGBoost dependencies published into Maven central by H2O
     // Versioning rules may differ for XGBoost artifacts published by H2O
