+++ h2o-genmodel/build.gradle
@@ -31,6 +31,5 @@ dependencies {
   compile("ai.h2o:deepwater-backend-api:1.0.4") {
     exclude group: 'org.slf4j'
   }
-  compile project(":h2o-genmodel-ext-xgboost")
   compileOnly "com.google.protobuf.nano:protobuf-javanano:3.1.0"
 }
